# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Wasl Executive Banking Portfolio Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List product integration records, optionally filtered by bank
      parameters:
        - name: bankId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product integration record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product integration record
      parameters:
        - $ref: "#/components/parameters/EntityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product integration record
      parameters:
        - $ref: "#/components/parameters/EntityId"
      responses:
        "204":
          description: Deleted
...
  /risks:
    get:
      operationId: listRisks
      tags: [risks]
      summary: List risks, optionally filtered by bank
      parameters:
        - name: bankId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of risks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Risk"
    post:
      operationId: createRisk
      tags: [risks]
      summary: Create a risk record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
  /risks/{id}:
    patch:
      operationId: updateRisk
      tags: [risks]
      summary: Update a risk record
      parameters:
        - $ref: "#/components/parameters/EntityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
    delete:
      operationId: deleteRisk
      tags: [risks]
      summary: Delete a risk record
      parameters:
        - $ref: "#/components/parameters/EntityId"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
...
    EntityId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Product:
      type: object
      required: [id, bankId, productCode, categoryStage, status, progressPercent, createdAt, updatedAt]
      properties:
        id: { type: integer }
        bankId: { type: string }
        productCode: { type: string }
        categoryStage: { type: string }
        status: { type: string }
        progressPercent: { type: number }
        dateType: { type: ["string", "null"] }
        dateValue: { type: ["string", "null"] }
        responsiblePerson: { type: ["string", "null"] }
        priorityImpact: { type: ["string", "null"] }
        descriptionNotes: { type: ["string", "null"] }
        riskLevel: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProductInput:
      type: object
      required: [bankId, productCode, categoryStage, status, progressPercent]
      properties:
        bankId: { type: string }
        productCode: { type: string }
        categoryStage: { type: string }
        status: { type: string }
        progressPercent: { type: number }
        dateType: { type: string }
        dateValue: { type: string }
        responsiblePerson: { type: string }
        priorityImpact: { type: string }
        descriptionNotes: { type: string }
        riskLevel: { type: string }
    ProductUpdate:
      type: object
      properties:
        productCode: { type: string }
        categoryStage: { type: string }
        status: { type: string }
        progressPercent: { type: number }
        dateType: { type: string }
        dateValue: { type: string }
        responsiblePerson: { type: string }
        priorityImpact: { type: string }
        descriptionNotes: { type: string }
        riskLevel: { type: string }
...
    Risk:
      type: object
      required: [id, bankId, description, level, status, createdAt]
      properties:
        id: { type: integer }
        bankId: { type: string }
        description: { type: string }
        level: { type: string }
        status: { type: string }
        createdAt: { type: string }
    RiskInput:
      type: object
      required: [bankId, description, level, status]
      properties:
        bankId: { type: string }
        description: { type: string }
        level: { type: string }
        status: { type: string }
    RiskUpdate:
      type: object
      properties:
        description: { type: string }
        level: { type: string }
        status: { type: string }
...